fixed materials
materials broke, along with the ability to remove blocks. readded the materials, fixed the cube's script

+++ Dwarven Core/Assets/My assets/Scripts/Inventory.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Inventory : MonoBehaviour
 {
     // Start is called before the first frame update
-    public int money = 0;
+    public float money = 0.00f;
     void Start()
     {
